+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [CHANGE] also recognize urls with scheme git+... as git
 * [ENHANCEMENT] tests can now run against a real docker host, improving test depth
 * [ENHANCEMENT] removed OsDb. This file contained some hardcoded Os versions which was not future proof. It was a workaround to get started more quickly.
+* [FIX] patched sources now fail when the underlying patch fails
 
 # 0.2.2 / 2016.09.05